Changed int to float

+++ src/main/java/rocks/zipcode/atm/CashMachine.java
@@ -33,7 +33,7 @@ public class CashMachine {
         }
     }
 
-    public Integer getBalance(int id) {
+    public Float getBalance(int id) {
         return bank.getBalanceOfAccount(id);
     }
 
@@ -52,7 +52,7 @@ public class CashMachine {
         );
     }
 
-    public void deposit(int amount) {
+    public void deposit(float amount) {
         if (accountData != null) {
             tryCall(
                     () -> bank.deposit(accountData, amount),
@@ -61,7 +61,7 @@ public class CashMachine {
         }
     }
 
-    public void withdraw(int amount) {
+    public void withdraw(float amount) {
         if (accountData != null) {
             tryCall(
                     () -> bank.withdraw(accountData, amount),
